Generate two-stage agg for scalar DQA on distribution column
Previously, for a query containing a scalar DQA(Distinct Qualified Aggregate)
on the distribution column, we never generated a plan alternative with a
two-stage aggregate. A scalar DQA is a DQA over a single group (most often
resulting from a query with no specified GROUP BY clause).

With this commit, the SplitDQA transform `CXformSplitDQA::Transform`, will
produce a two-stage aggregate alternative for such a query.

Signed-off-by: Ekta Khanna <ekhanna@pivotal.io>