Add mock of CDrvdPropRelational
Because we make new derived relational properties for an expression (as
opposed to for a group expression where these should already exist [I
think]), we need to, for now, make a mock of this in order to avoid
having to make a real one, which would then entail us making real
child expressions with real operators and all of the associated members
which are filled into their derived properties in the Derive function
call which gets called on the CDrvdPropRelational object.

The next steps as far as I can see are to make the derived properties in
the ExpressionHandleMock's DeriveProps and see if setting the list of
derived properties in the expressionalhandlemock's member variable will
allow us to then call PdpDerive on the mock expression and get back the
desired colrefset. We will need to define PcrsOuter on our
CDrvdPropRelational.

We also have to mock the CDrvdPropRelational and not just hollow out the
call to PdpDerive in CExpression because we call PcrsOuter on the
CDrvdPropRelational and we want to be able to replace that in our mock
code path.

The question arises, why not simply make the CDrvdPropRelationalMock
inside of CExpressionMock::PdpDerive instead of bothering to make a mock
CExpressionHandle and then make a CExpressionHandleMock::DeriveProps and
do it there? That is a very good question which I cannot answer. We can
go back and try that, however, for now, I made the expressionhandlemock,
so let's use it and keep PdpDerive similar to the real one. Without the
expressionhandlemock, we would only have derived properties for the
expression and not for any children, as those are stored in the
expressionhandle. Though, in this case, that does not seem to strictly
matter.
After getting this to work, maybe try doing this without any expression
handle mock.