Move logic for returning relational prop into CExpressionMock
Move the logic for returning derived properties from the expression
handle mock into the expression mock. This gives us more control.
However, now we have the problem that we need the relational properties
returned and set in the mock expression to be the mock relational
properties. (it makes sense that we made them because anything your mock
makes, you should make and control) However, in the production code, it
calls a cast function on the returned derived properties that were
created by the mock cexpression which casts the return to a
CDrvdPropRelational. We need it to be a CDrvdPropRelationalMock so we
can call our PcrsOuter on it. I need to investigate more if I can do
this with how I define the cast function (since it is static, this might
be harder) or if I need to change the production code. Also I need to
know the intent behind manually force casting this (is it the error
message?).