Mock out CDrvdRelationalProperties
- Mock our derived relational properties
- Do super hacky stuff to compile in symbols for CTestUtils

In order to create CColRefSet, we need to mock out
CDrvdRelationalProperties for which I need access to CTestUtils
TestSetup methods. This means I had to hack into the libgpopt
CMakeLists.txt to be able to get access to the unittest symbols

Also, hack Pdprel to get it to not cast our mock relational properties
to real ones. Also, hack the unittest array to hardcode it to only run
our test because some other stuff is broken.

However, I finally got my colrefset to be returned. This was really
painful. I need to figure out where to put the mock code.
Next steps include clean up and mocking out the overloaded [] operator
for CExpression so that I get a fake kid. I think.