Correct colrefset and outerrefs
Correct which columns are grouping columns and which are outerrefs and
which are the minimal FDs.
Also, make a new PcrCreate() in ColRefFactory which makes it so that I
can specify the col_id. This makes it so I can get them to match up.
THere are definitely other/better ways of doing this, but it is
temporary.

We now have a correct reduced colrefset coming out of the exclude
columns. However, we need to figure out how to handle recursion.
In order to recursively process children and to meet the condition that
we have a non-empty project list, we will have to solve the recursion
problem.
How to recursively process children without specifying all of them?