Hardcode in counter for constant grp grpexpr
Add relevant counters to a particular xform, RemoveSubqDistinct, which
are hard-coded to only increment for the same group number and group
expression. It is meant to show how many times we extract from the memo
an expression rooted at this group expression in this group and how many
times it matches the pattern and we apply the xform and how many times
it produces an alternative. It isn't working, however. It seems like the
counters are getting reset to zero, because with the debugger I saw it
should be at least 2.